+++ pkg/main.go
@@ -16,7 +16,8 @@ func main() {
 	// argument. This factory will be automatically called on incoming request
 	// from Grafana to create different instances of `App` (per plugin
 	// ID).
-	if err := app.Manage(plugin.Name, plugin.NewDashboardReporterApp, app.ManageOpts{}); err != nil {
+	err := app.Manage(plugin.Name, plugin.NewDashboardReporterApp, app.ManageOpts{})
+	if err != nil {
 		log.DefaultLogger.Error("failed to start "+plugin.Name, "err", err.Error())
 		os.Exit(1)
 	}